+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+3.6.0 (xxxx-xx-xx)
+--
+Improvements:
+
+* Core: introduction of the Console component.
+* Console: added command which prepares for a reinstall.
+
+
 3.5.1 (xxxx-xx-xx)
 --
 Improvements: